+++ src/app/integrations/integrations.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { AuthService } from 'app/core/auth.service';
 import { IntegrationService } from 'app/services/integration.service';
-import { APPS_TITLE, BrevoIntegration, N8nIntegration, CATEGORIES_LIST, CustomerioIntegration, HubspotIntegration, INTEGRATIONS_CATEGORIES, INTEGRATIONS_KEYS, INTEGRATION_LIST_ARRAY, MakeIntegration, OpenaiIntegration, QaplaIntegration, INTEGRATION_LIST_ARRAY_CLONE } from './utils';
+import { APPS_TITLE, BrevoIntegration, N8nIntegration, CATEGORIES_LIST, CustomerioIntegration, HubspotIntegration, INTEGRATIONS_CATEGORIES, INTEGRATIONS_KEYS, INTEGRATION_LIST_ARRAY, MakeIntegration, OpenaiIntegration, QaplaIntegration, INTEGRATION_LIST_ARRAY_CLONE, GoogleIntegration, AnthropicIntegration, GroqIntegration, CohereIntegration,  } from './utils'; // OllamaIntegration, DeepseekIntegration
 import { LoggerService } from 'app/services/logger/logger.service';
 import { NotifyService } from 'app/core/notify.service';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -544,6 +544,31 @@ export class IntegrationsComponent implements OnInit, OnDestroy {
     if (key === INTEGRATIONS_KEYS.OPENAI) {
       return new OpenaiIntegration();
     }
+
+    if (key === INTEGRATIONS_KEYS.GOOGLE) {
+      return new GoogleIntegration();
+    }
+
+    if (key === INTEGRATIONS_KEYS.ANTHROPIC) {
+      return new AnthropicIntegration();
+    }
+
+    if (key === INTEGRATIONS_KEYS.GROQ) {
+      return new GroqIntegration();
+    }
+
+    if (key === INTEGRATIONS_KEYS.COHERE) {
+      return new CohereIntegration();
+    }
+
+    // if (key === INTEGRATIONS_KEYS.OLLAMA) {
+    //   return new OllamaIntegration();
+    // }
+
+    // if (key === INTEGRATIONS_KEYS.DEEPSEEK) {
+    //   return new DeepseekIntegration();
+    // }
+
     if (key === INTEGRATIONS_KEYS.MAKE) {
       return new MakeIntegration();
     }