+++ src/app/integrations/list/groq-integration/groq-integration.component.ts
@@ -1,32 +1,31 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { IntegrationService } from 'app/services/integration.service';
 import { LoggerService } from 'app/services/logger/logger.service';
-import { OpenaiService } from 'app/services/openai.service';
 
 @Component({
-  selector: 'openai-integration',
-  templateUrl: './openai-integration.component.html',
-  styleUrls: ['./openai-integration.component.scss']
+  selector: 'groq-integration',
+  templateUrl: './groq-integration.component.html',
+  styleUrls: ['./groq-integration.component.scss']
 })
-export class OpenaiIntegrationComponent implements OnInit {
+export class GroqIntegrationComponent implements OnInit {
 
   @Input() integration: any;
-  @Output() onUpdateIntegration = new EventEmitter;
-  @Output() onDeleteIntegration = new EventEmitter;
+    @Output() onUpdateIntegration = new EventEmitter;
+    @Output() onDeleteIntegration = new EventEmitter;
 
-  keyVisibile: boolean = false;
+    keyVisibile: boolean = false;
   isVerified: boolean;
   translateparams: any;
   isMasked: boolean = true; // State for masking
 
   constructor(
-    private integrationService: IntegrationService,
-    private logger: LoggerService
+       private integrationService: IntegrationService,
+        private logger: LoggerService
   ) { }
 
   ngOnInit(): void {
-    this.logger.log("[INT-OpenAI] integration ", this.integration)
-    this.translateparams = { intname: 'OpenAI' };
+    this.logger.log("[INT-GROQ] integration ", this.integration)
+    this.translateparams = { intname: 'Groq' };
     if (this.integration.value.apikey) {
       this.checkKey();
     }
@@ -59,13 +58,14 @@ export class OpenaiIntegrationComponent implements OnInit {
 
   checkKey() {
     return new Promise((resolve) => {
-      let url = "https://api.openai.com/v1/models";
+      let url = "https://api.groq.com/openai/v1/models";
       let key = "Bearer " + this.integration.value.apikey;
       this.integrationService.checkIntegrationKeyValidity(url, key).subscribe((resp) => {
+        this.logger.log("[INT-GROQ] Key verification resp: ", resp);
         this.isVerified = true;
         resolve(true);
       }, (error) => {
-        this.logger.error("[INT-OpenAI] Key verification failed: ", error);
+        this.logger.error("[INT-GROQ] Key verification failed: ", error);
         this.isVerified = false;
         resolve(false);
       })
@@ -79,7 +79,7 @@ export class OpenaiIntegrationComponent implements OnInit {
     }
   }
 
-  // ---------------------------------------------------
+    // ---------------------------------------------------
   // Mask Api key without use input of password type
   // ---------------------------------------------------
   handleInput(event: Event): void {