+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # tiledesk-dashboard
 
+### 2.7.75
+- Makes AI integrations available from the Premium plan
+
+### 2.7.74
+- Adds Google Gemini integration
+- Adds Claude integration
+- Adds Croq integration
+- Adds Cohere integration
+
 ### 2.7.73
 - Fixes visibility of blocks in Reassign Conversation